# add tx-ingest-benchmark, cut out O(n²) operation in DUV.splitAndTransfer