Adds a benchmark that simply ingests 100k simple docs (`{:xt/id n}`) in various batch sizes, and fixes some low-hanging fruit in Arrow's DUV.

Before (my machine):

- batches of 1000: 180ms
- batches of 100: 1092ms
- batches of 10: 3177ms
- individual: 26s

![before profile](https://github.com/user-attachments/assets/4b5253c8-38e0-43d6-b62e-b3b9997f8d36)

After:

- batches of 1000: 117ms
- batches of 100: 333ms
- batches of 10: 954ms
- individual: 6.5s

![after profile](https://github.com/user-attachments/assets/9aafbca5-8d31-4cd0-8713-c33992c9e3a9)

The performance improvement is a fast-path in DUV.splitAndTransfer when the transfer is for the whole vector - we don't need to scan the vector to create the target vector's offset buffer, because it'll be the same as the source vector.

Next up on this particular path would likely be trying to remove the necessity of eager watermarks introduced in #3335 (introduced for correctness, but took a performance hit).

(hopefully goes without saying, users should still try to batch puts wherever possible)